refactor(convention/plugins/extensions/*): refactored
