+++ build-logic/convention/src/main/kotlin/corp/tbm/cleanwizard/buildLogic/convention/plugins/extensions/CleanWizardCodegenExtension.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanwizard.buildLogic.convention.processorConfig
+package corp.tbm.cleanwizard.buildLogic.convention.plugins.extensions
 
 open class CleanWizardCodegenExtension(
     var domainProject: String = "",
